backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.4
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.24.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
